+++ packages/profile-metrics-controller/src/ProfileMetricsService.ts
@@ -6,6 +6,7 @@ import { createServicePolicy, HttpError } from '@metamask/controller-utils';
 import type { Messenger } from '@metamask/messenger';
 import { SDK } from '@metamask/profile-sync-controller';
 import type { AuthenticationController } from '@metamask/profile-sync-controller';
+import type { IDisposable } from 'cockatiel';
 
 import type { ProfileMetricsServiceMethodActions } from '.';
 
@@ -150,7 +151,7 @@ export class ProfileMetricsService {
    * {@link CockatielEvent}.
    * @see {@link createServicePolicy}
    */
-  onRetry(listener: Parameters<ServicePolicy['onRetry']>[0]) {
+  onRetry(listener: Parameters<ServicePolicy['onRetry']>[0]): IDisposable {
     return this.#policy.onRetry(listener);
   }
 
@@ -163,7 +164,7 @@ export class ProfileMetricsService {
    * {@link CockatielEvent}.
    * @see {@link createServicePolicy}
    */
-  onBreak(listener: Parameters<ServicePolicy['onBreak']>[0]) {
+  onBreak(listener: Parameters<ServicePolicy['onBreak']>[0]): IDisposable {
     return this.#policy.onBreak(listener);
   }
 
@@ -184,7 +185,9 @@ export class ProfileMetricsService {
    * @returns An object that can be used to unregister the handler. See
    * {@link CockatielEvent}.
    */
-  onDegraded(listener: Parameters<ServicePolicy['onDegraded']>[0]) {
+  onDegraded(
+    listener: Parameters<ServicePolicy['onDegraded']>[0],
+  ): IDisposable {
     return this.#policy.onDegraded(listener);
   }
 
@@ -197,7 +200,7 @@ export class ProfileMetricsService {
   async submitMetrics(data: ProfileMetricsSubmitMetricsRequest): Promise<void> {
     const authToken = await this.#messenger.call(
       'AuthenticationController:getBearerToken',
-      data.entropySourceId || undefined,
+      data.entropySourceId ?? undefined,
     );
     await this.#policy.execute(async () => {
       const url = new URL(`${this.#baseURL}/profile/accounts`);